+++ src/main/java/com/pretallez/controller/VotePostController.java
@@ -1,15 +1,17 @@
 package com.pretallez.controller;
 
-import com.pretallez.common.response.CustomApiResponse;
-import com.pretallez.common.response.ResSuccessCode;
-import com.pretallez.votepost.model.VotePostCreate;
-import com.pretallez.votepost.service.VotePostService;
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.pretallez.common.response.CustomApiResponse;
+import com.pretallez.common.response.ResSuccessCode;
+import com.pretallez.domain.votepost.dto.VotePostCreate;
+import com.pretallez.domain.votepost.service.VotePostService;
+
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
 
 @RestController
 @RequiredArgsConstructor
