+++ src/test/java/com/pretallez/integration/service/BoardServiceImplIntegrationTest.java
@@ -5,8 +5,9 @@ import com.pretallez.common.fixture.MemberFixture;
 import com.pretallez.domain.board.dto.BoardCreate;
 import com.pretallez.common.entity.Board;
 import com.pretallez.common.entity.Member;
-import com.pretallez.member.repository.MemberRepository;
 import com.pretallez.domain.board.service.BoardService;
+import com.pretallez.domain.member.repository.MemberRepository;
+
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
