+++ src/test/java/com/pretallez/integration/service/ChatroomServiceIntegrationTest.java
@@ -3,16 +3,16 @@ package com.pretallez.integration.service;
 import com.pretallez.domain.board.repository.BoardRepository;
 import com.pretallez.domain.chatroom.repository.ChatroomRepository;
 import com.pretallez.common.fixture.*;
-import com.pretallez.fencingclub.repository.FencingClubRepository;
-import com.pretallez.member.repository.MemberRepository;
 import com.pretallez.domain.chatroom.dto.ChatroomCreate;
 import com.pretallez.common.entity.Board;
 import com.pretallez.common.entity.FencingClub;
 import com.pretallez.common.entity.Member;
 import com.pretallez.common.entity.VotePost;
-import com.pretallez.votepost.service.VotePostService;
 import com.pretallez.domain.chatroom.service.ChatroomServiceImpl;
-import com.pretallez.votepost.repository.VotePostRepository;
+import com.pretallez.domain.fencingclub.repository.FencingClubRepository;
+import com.pretallez.domain.member.repository.MemberRepository;
+import com.pretallez.domain.votepost.repository.VotePostRepository;
+import com.pretallez.domain.votepost.service.VotePostService;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
