+++ src/test/java/com/pretallez/integration/service/VotePostServiceImplIntegrationTest.java
@@ -3,15 +3,16 @@ package com.pretallez.integration.service;
 import com.pretallez.common.fixture.FencingClubFixture;
 import com.pretallez.common.fixture.MemberFixture;
 import com.pretallez.common.fixture.VotePostFixture;
-import com.pretallez.votepost.model.VotePostCreate;
 import com.pretallez.common.entity.Board;
 import com.pretallez.common.entity.FencingClub;
 import com.pretallez.common.entity.Member;
 import com.pretallez.common.entity.VotePost;
-import com.pretallez.fencingclub.repository.FencingClubRepository;
-import com.pretallez.member.repository.MemberRepository;
-import com.pretallez.votepost.repository.VotePostRepository;
-import com.pretallez.votepost.service.VotePostService;
+import com.pretallez.domain.fencingclub.repository.FencingClubRepository;
+import com.pretallez.domain.member.repository.MemberRepository;
+import com.pretallez.domain.votepost.dto.VotePostCreate;
+import com.pretallez.domain.votepost.repository.VotePostRepository;
+import com.pretallez.domain.votepost.service.VotePostService;
+
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
