+++ src/test/java/com/pretallez/unit/controller/MemberChatroomControllerUnitTest.java
@@ -23,11 +23,11 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.pretallez.common.fixture.MemberChatroomFixture;
 import com.pretallez.common.response.ResSuccessCode;
 import com.pretallez.controller.MemberChatroomController;
-import com.pretallez.memberchatroom.model.ChatroomMembersRead;
-import com.pretallez.memberchatroom.model.MemberChatroomCreate;
-import com.pretallez.memberchatroom.model.MemberChatroomDelete;
-import com.pretallez.memberchatroom.model.MemberChatroomsRead;
-import com.pretallez.memberchatroom.service.MemberChatroomService;
+import com.pretallez.domain.memberchatroom.dto.ChatroomMembersRead;
+import com.pretallez.domain.memberchatroom.dto.MemberChatroomCreate;
+import com.pretallez.domain.memberchatroom.dto.MemberChatroomDelete;
+import com.pretallez.domain.memberchatroom.dto.MemberChatroomsRead;
+import com.pretallez.domain.memberchatroom.service.MemberChatroomService;
 
 @ExtendWith(MockitoExtension.class)
 @DisplayName("회원 채팅방 컨트롤러 단위 테스트")
